pyproject.toml
[project]
name = "nlp"
version = "0.1.0"
description = ""
authors = [{ name = "Tatsuki Okada", email = "tatsukio0522@gmali.com" }]
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.0",
    "setuptools>=69.0.3",
    "fire>=0.5.0",
    "pydantic>=2.5.3",
    "beautifulsoup4>=4.12.2",
    "selenium>=4.16.0",
    "fastapi>=0.108.0",
    "uvicorn>=0.25.0",
    "matplotlib>=3.9.0",
    "pandas>=1.4.2",
    "seaborn>=0.11.2",
    "japanize-matplotlib>=1.1.3",
    "numpy>=2.2.0",
    "tqdm>=4.64.0",
    "scikit-learn>=1.6.0",
    "openpyxl>=3.1.2",
    "polars>=1.29.0",
    "aiohttp>=3.9.5",
    "tenacity>=8.4.1",
    "playwright>=1.44.0",
    "toml>=0.10.2",
    "sentencepiece>=0.2.0",
    "protobuf>=3.20.1",
    "transformers>=4.51.0",
    "spacy>=3.3.0",
    "mecab-python3>=1.0.4",
    "datasets>=3.5.0",
    "demoji>=1.1.0",
    "emoji>=1.7.0",
    "mojimoji>=0.0.12",
    "evaluate>=0.4.3",
    "accelerate>=1.6.0",
    "pytorch-lightning>=2.4.0",
    "optimum>=1.13.2",
    "pillow>=10.3.0",
    "trl>=0.17.0",
    "langchain>=0.2.5",
    "optuna>=4.3.0",
    "peft>=0.15.2",
    "litellm>=1.67.0",
    "torch==2.6.0+cu126; sys_platform == 'linux' and platform_machine == 'x86_64'",
    "torch==2.6.0; sys_platform == 'darwin' or (sys_platform == 'linux' and platform_machine == 'aarch64')",
    "haystack>=0.42",
    "gensim>=4.3.2",
    "hatchling>=1.27.0",
    "editables>=0.5",
    "wheel>=0.45.1",
    "sentence-transformers>=4.1.0",
    "datasketch>=1.6.5",
    "bm25s>=0.2.12",
    "jsonlines>=4.0.0",
    "sudachipy>=0.6.10",
    "sudachidict-core>=20250129",
]

[tool.hatch.build.targets.wheel]
packages = ["src/nlp", "src"]

[tool.uv]
dev-dependencies = [
    "tensorboard>=2.14.0",
    "wandb>=0.19.10",
    "deepspeed>=0.16.7",
    "jupyterlab>=3.3.4",
    "marimo>=0.13.3",
    "mypy>=1.11.2",
    "nbstripout>=0.7.1",
    "pre-commit>=4.0.1",
    "pytest>=8.3.3",
    "ruff>=0.6.9",
    "tox>=4.21.2",
    "types-pyyaml>=6.0.12.20250402",
    "types-toml>=0.10.8.20240310",
    "pytest-asyncio>=0.26.0",
]

[[tool.uv.index]]
name = "torch-cuda"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[[tool.uv.index]]
name = "torch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = [
    { index = "torch-cuda", marker = "sys_platform == 'linux' and platform_machine == 'x86_64'" },
    { index = "torch-cpu", marker = "sys_platform == 'darwin' or (sys_platform == 'linux' and platform_machine == 'aarch64')" },
]


[tool.ruff]
line-length = 119
target-version = "py312"
exclude = [".git", ".venv", "__pycache__", "data", "dist", "misc", "notebooks", "prof", "tmp", "workspacea", ".tox"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
extend-select  = [
    "I", # isort
]
# ignores
ignore = ["E501"]

[tool.ruff.lint.isort]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
split-on-trailing-comma = true

[tool.ruff.lint.pyupgrade]
# Python3.8互換のための設定
keep-runtime-typing = true

[tool.mypy]
python_version="3.12"
files = "src"
ignore_missing_imports = true
disallow_untyped_defs = true
no_implicit_optional = true
allow_redefinition = true
show_error_codes = true
pretty = true
allow_untyped_globals = true

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning",]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
